refactoring and debugging

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -6,14 +6,12 @@ use App\Http\Filters\V1\AuthorFilter;
 use App\Http\Requests\Api\V1\ReplaceUserRequest;
 use App\Http\Requests\Api\V1\StoreUserRequest;
 use App\Http\Requests\Api\V1\UpdateUserRequest;
-use App\Http\Resources\V1\TicketResource;
 use App\Http\Resources\V1\UserResource;
-use App\Models\Ticket;
 use App\Models\user;
 use App\Policies\V1\UserPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Auth;
+
 
 class UserController extends ApiController
 {
@@ -33,19 +31,17 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            $this->isAble('store', User::class);
+        if ($request->user()->is_manager)
             return new UserResource(User::create($request->mappedAttributes()));
-        } catch (AuthorizationException $e) {
-            Log::error('User creation failed: ' . $e->getMessage());
-            return $this->error("You're not authorized to create that resource"  . $e->getMessage(), 401);
-        }
+        else return $this->error('You are not authorized to update that resource', 403);
+
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(User $user)
+    public
+    function show(User $user)
     {
         if ($this->include('tickets')) {
             return new UserResource($user->load('tickets'));
@@ -56,7 +52,8 @@ class UserController extends ApiController
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(User $user)
+    public
+    function edit(User $user)
     {
         //
     }
@@ -64,55 +61,52 @@ class UserController extends ApiController
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateUserRequest $request, $user_id)
+    public
+    function update(UpdateUserRequest $request, $user_id)
     {
         try {
             $user = User::findOrFail($user_id);
-          //  Log::info('Received user_id:', ['user_id' => $user_id]);
+            //  Log::info('Received user_id:', ['user_id' => $user_id]);
 
             //Policy
-            $this->isAble('update', $user);
-            $user->update($request->mappedAttributes());
-            return new UserResource($user);
-
+            if ($request->user()->is_manager) {
+                $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            } else return $this->error('You are not authorized to update that resource', 403);
         } catch (ModelNotFoundException) {
             return $this->error('User cannot be found', 404);
-        } catch (AuthorizationException) {
-            return $this->error('You are not authorized to update that resource', 403);
         }
-
     }
 
-    public function replace(ReplaceUserRequest $request, $user_id)
+    public
+    function replace(ReplaceUserRequest $request, $user_id)
     {
         try {
             $user = User::findOrFail($user_id);
 
             //Policy
-            $this->isAble('replace', $user);
-            $user->update($request->mappedAttributes());
-
-            return new UserResource($user);
-
+            if ($request->user()->is_manager) {
+                $this->isAble('replace', $user);
+                $user->update($request->mappedAttributes());
+                return new UserResource($user);
+            }
         } catch (ModelNotFoundException) {
             return $this->error('User cannot be found', 404);
-        } catch (AuthorizationException) {
-            return $this->error('You are not authorized to replace that resource', 403);
         }
-
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($user_id)
+    public
+    function destroy($user_id)
     {
         try {
             $user = User::findOrFail($user_id);
-            $this->isAble('delete', $user);
-            $user->delete();
-
-            return $this->ok("user successfully deleted");
+            if (Auth::user()->is_manager) {
+                $user->delete();
+                return $this->ok("user successfully deleted");
+            } else $this->error('User cannot be found', 404);
         } catch (ModelNotFoundException) {
             return $this->error('user cannot be found', 404);
         }
